Extract media editor save hook
